refactor(event-emitter)
