+++ src/base/static/components/molecules/form-field-types/geolocate-field.js
@@ -5,7 +5,7 @@ import styled from "@emotion/styled";
 import { Button } from "../../atoms/buttons";
 import { FontAwesomeIcon } from "../../atoms/imagery";
 
-import emitter from "../../../utils/emitter";
+import emitter from "../../../utils/event-emitter";
 import constants from "../../../constants";
 
 const GeolocateButton = styled(props => (
