--- src/base/static/utils/emitter.js
@@ -1,3 +1,0 @@
-import { EventEmitter } from "fbemitter";
-
-export default new EventEmitter();
